# Conflicts:
#	app/src/main/java/com/uguinformatica/bluemoon/androidapp/MainActivity.kt
#	app/src/main/java/com/uguinformatica/bluemoon/androidapp/domain/models/Tradeable.kt
#	app/src/main/java/com/uguinformatica/bluemoon/androidapp/ui/components/OrderComponents/OrderItem.kt
#	app/src/main/java/com/uguinformatica/bluemoon/androidapp/ui/components/ProductComponents/ProductListItem.kt
#	app/src/main/java/com/uguinformatica/bluemoon/androidapp/ui/components/UserDataComponents/ModifyButton.kt
#	app/src/main/java/com/uguinformatica/bluemoon/androidapp/ui/screens/UserDataScreen.kt

+++ app/src/main/java/com/uguinformatica/bluemoon/androidapp/domain/models/SilverType.kt
@@ -1,3 +1,3 @@
 package com.uguinformatica.bluemoon.androidapp.domain.models
 
-data class SilverType(var name: String, var currentPrice: Float)
+data class SilverType(var name: String, var currentPrice: Double)
